backend/requirements.txt
absl-py==1.3.0
aio-pika==8.3.0
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.3
aiormq==6.6.4
aiosignal==1.3.1
altgraph==0.17.3
APScheduler==3.9.1.post1
apturl==0.5.2
asgiref==3.7.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
auto-py-to-exe==2.27.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.12.0
betterpath==0.2.2
bidict==0.22.1
blinker==1.4
boto3==1.26.101
botocore==1.29.101
bottle==0.12.23
bottle-websocket==0.2.9
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
channels==4.0.0
chardet==4.0.0
charset-normalizer==2.1.1
chatbotAI==0.3.1.3
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.4
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==1.9.2
crispy-bootstrap4==2022.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dask==2022.10.2
dbus-python==1.2.18
defer==1.0.6
discord==2.1.0
discord.py==2.1.0
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.4
django-cors-headers==4.2.0
django-crispy-forms==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.21
Eel==0.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fasteners==0.14.1
fbmessenger==6.0.0
filelock==3.12.3
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.2
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-images-download==2.8.0
Google-Images-Search==1.4.6
google-pasta==0.2.0
googleapis-common-protos==1.59.0
googlesearch-python==1.2.3
gravityai==0.1.3.post1
greenlet==2.0.1
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httplib2==0.20.2
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==4.6.4
jax==0.4.6
jeepney==0.7.1
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
l==0.11.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mattermostwrapper==2.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
msgpack==1.0.5
multidict==5.2.0
networkx==2.6.3
nltk==3.8.1
noise==1.2.2
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
openai==0.25.0
opencv-python-headless==4.7.0.68
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==20.9
pamqp==3.2.1
pandas==1.5.2
pandas-stubs==1.5.2.221124
paramiko==2.9.3
partd==1.3.0
pathlib==1.0.1
pexpect==4.8.0
Pillow==9.4.0
pipenv==2023.8.28
platformdirs==3.10.0
playsound==1.3.0
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==4.22.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.2
pydot==1.4.2
pyfiglet==0.8.post1
pygame==2.5.2
PyGetWindow==0.0.9
PyGObject==3.42.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyJWT==2.3.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==3.10.1
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-crfsuite==0.9.9
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-engineio==4.4.0
python-resize-image==1.1.20
python-socketio==5.8.0
python3-xlib==0.15
PyTMX==3.31
pytube==12.1.3
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyvoicechanger==1.5.0
pyxdg==0.27
PyYAML==5.4.1
questionary==1.10.0
randomname==0.1.5
rasa==3.5.1
rasa-sdk==3.5.0
redis==4.5.3
regex==2022.10.31
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
SecretStorage==3.3.1
selenium==4.9.0
sentry-sdk==1.14.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.20.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.4.47
sqlparse==0.4.4
systemd-python==234
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.11.0
termcolor==1.1.0
terminaltables==3.1.10
tflearn==0.5.0
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.2
twilio==7.14.2
typeguard==3.0.2
types-pytz==2022.6.0.1
typing-utils==0.1.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.15
usb-creator==0.3.7
uvloop==0.17.0
vcversioner==2.16.0.0
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.3
whichcraft==0.6.1
whitenoise==6.7.0
wikipedia==1.4.0
Wikipedia-API==0.5.8
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
zope.event==4.6
zope.interface==5.5.2
